Query the ClinicalTrials.gov dataset

This function selects a subset of the clinical trials data by using a a variety of different search parameters. These include free text search keywords, range queries for the continuous variables, and exact matches for categorical fields. The function ctgov_query_terms shows the categorical levels for the latter. The function will either take the entire dataset loaded into the package environment or a previously queried input.

ctgov_query(
  data = NULL,
  description_kw = NULL,
  source_kw = NULL,
  brief_title_kw = NULL,
  official_title_kw = NULL,
  criteria_kw = NULL,
  intervention_desc_kw = NULL,
  conditions_kw = NULL,
  population_kw = NULL,
  date_range = NULL,
  enrollment_range = NULL,
  minimum_age_range = NULL,
  maximum_age_range = NULL,
  study_type = NULL,
  allocation = NULL,
  intervention_model = NULL,
  observational_model = NULL,
  primary_purpose = NULL,
  time_perspective = NULL,
  masking_description = NULL,
  sampling_method = NULL,
  phase = NULL,
  gender = NULL,
  sponsor_type = NULL,
  ignore_case = TRUE,
  match_all = FALSE
)

Arguments

data
a dataset to search over; set to NULL to use the full dataset that is currently loaded

description_kw
character vector of keywords to search in the intervention description field. Set to NULL to avoid searching this field.

source_kw
character vector of keywords to search in the source (the company that submitted the study). Set to NULL to avoid searching this field.

brief_title_kw
character vector of keywords to search in the brief title field. Set to NULL to avoid searching this field.

official_title_kw
character vector of keywords to search in the official title field. Set to NULL to avoid searching this field.

criteria_kw
character vector of keywords to search in the criteria field. Set to NULL to avoid searching this field.

intervention_desc_kw
character vector of keywords to search in the intervention description field. Set to NULL to avoid searching this field.

conditions_kw
character vector of keywords to search in the conditions field. Set to NULL to avoid searching this field.

population_kw
character vector of keywords to search in the population field. Set to NULL to avoid searching this field.

date_range
string of length two formatted as "YYYY-MM-DD" describing the earliest and latest data to include in the results. Use a missing value for either value search all dates. Set to NULL to avoid searching this field.

enrollment_range
numeric of length two describing the smallest and largest enrollment sizes to include in the results. Use a missing value for either value to avoid filtering. Set to NULL to avoid searching this field.

minimum_age_range
numeric of length two describing the smallest and largest minmum age (in years) to include in the results. Use a missing value for either value to avoid filtering. Set to NULL to avoid searching this field.

maximum_age_range
numeric of length two describing the smallest and largest maximum age (in years) to include in the results. Use a missing value for either value to avoid filtering. Set to NULL to avoid searching this field.

study_type
character vector of study types to include in the output. Set to NULL to avoid searching this field.

allocation
character vector of allocations to include in the output. Set to NULL to avoid searching this field.

intervention_model
character vector of interventions to include in the output. Set to NULL to avoid searching this field.

observational_model
character vector of observations to include in the output. Set to NULL to avoid searching this field.

primary_purpose
character vector of primary purposes to include in the output. Set to NULL to avoid searching this field.

time_perspective
character vector of time perspectives to include in the output. Set to NULL to avoid searching this field.

masking_description
character vector of maskings to include in the output. Set to NULL to avoid searching this field.

sampling_method
character vector of sampling methods to include in the output. Set to NULL to avoid searching this field.

phase
character vector of phases to include in the output. Set to NULL to avoid searching this field.

gender
character vector of genders to include in the output. Set to NULL to avoid searching this field.

sponsor_type
character vector of sponsor types to include in the output. Set to NULL to avoid searching this field.

ignore_case
logical. Should the search ignore capitalization. The default is TRUE.

match_all
logical. Should the results required matching all the keywords? The default is FALSE.

Author

Taylor B. Arnold, taylor.arnold@acm.org